chore: basic auth file rename
