+++ packages/pieces/src/lib/framework/property/index.ts
@@ -1,5 +1,5 @@
 export * from './base-prop';
-export * from './custom-auth-prop';
+export * from './basic-auth-prop';
 export * from './dropdown-prop';
 export * from './oauth2-prop';
 export * from './property';
